+++ test/e2e/btc_staking_e2e_test.go
@@ -311,7 +311,7 @@ func (s *BTCStakingTestSuite) Test3CommitPublicRandomnessAndSubmitFinalitySignat
 	// ensure BTC delegation has received rewards after the block is finalised
 	btcDelRewardGauges, err := nonValidatorNode.QueryRewardGauge(delBabylonAddr)
 	s.NoError(err)
-	btcDelRewardGauge, ok := btcDelRewardGauges[itypes.BTC_DELEGATION.String()]
+	btcDelRewardGauge, ok := btcDelRewardGauges[itypes.BTC_STAKER.String()]
 	s.True(ok)
 	s.True(btcDelRewardGauge.Coins.IsAllPositive())
 	s.T().Logf("the finality provider received rewards for providing finality")
@@ -323,7 +323,7 @@ func (s *BTCStakingTestSuite) Test4WithdrawReward() {
 	s.NoError(err)
 
 	n.WithdrawRewardCheckingBalances(itypes.FINALITY_PROVIDER.String(), s.cacheFP.Addr)
-	n.WithdrawRewardCheckingBalances(itypes.BTC_DELEGATION.String(), s.cacheFP.Addr)
+	n.WithdrawRewardCheckingBalances(itypes.BTC_STAKER.String(), s.cacheFP.Addr)
 }
 
 // Test5SubmitStakerUnbonding is an end-to-end test for user unbonding
@@ -354,9 +354,20 @@ func (s *BTCStakingTestSuite) Test5SubmitStakerUnbonding() {
 	currentBtcTip, err := chain.ParseBTCHeaderInfoResponseToInfo(currentBtcTipResp)
 	s.NoError(err)
 
-	unbondingTx := activeDel.BtcUndelegation.UnbondingTx
-	unbondingTxMsg, err := bbn.NewBTCTxFromBytes(unbondingTx)
-	s.NoError(err)
+	unbondingTx := activeDel.MustGetUnbondingTx()
+
+	_, unbondingTxMsg := datagen.AddWitnessToUnbondingTx(
+		s.T(),
+		stakingMsgTx.TxOut[activeDel.StakingOutputIdx],
+		s.delBTCSK,
+		s.covenantSKs,
+		s.covenantQuorum,
+		[]*btcec.PublicKey{s.cacheFP.BtcPk.MustToBTCPK()},
+		uint16(activeDel.GetStakingTime()),
+		int64(activeDel.TotalSat),
+		unbondingTx,
+		s.net,
+	)
 
 	blockWithUnbondingTx := datagen.CreateBlockWithTransaction(s.r, currentBtcTip.Header.ToBlockHeader(), unbondingTxMsg)
 	nonValidatorNode.InsertHeader(&blockWithUnbondingTx.HeaderBytes)
@@ -368,6 +379,9 @@ func (s *BTCStakingTestSuite) Test5SubmitStakerUnbonding() {
 			&stakingTxHash,
 			unbondingTxMsg,
 			inclusionProof,
+			[]*wire.MsgTx{
+				stakingMsgTx,
+			},
 		)
 		// wait for a block so that above txs take effect
 		nonValidatorNode.WaitForNextBlock()
@@ -824,6 +838,7 @@ func ParseRespBTCDelToBTCDel(resp *bstypes.BTCDelegationResponse) (btcDel *bstyp
 		BtcPk:            resp.BtcPk,
 		FpBtcPkList:      resp.FpBtcPkList,
 		StartHeight:      resp.StartHeight,
+		StakingTime:      resp.StakingTime,
 		EndHeight:        resp.EndHeight,
 		TotalSat:         resp.TotalSat,
 		StakingTx:        stakingTx,