+++ x/btcstaking/keeper/msg_server.go
@@ -562,13 +562,37 @@ func (ms msgServer) AddCovenantSigs(goCtx context.Context, req *types.MsgAddCove
 	return &types.MsgAddCovenantSigsResponse{}, nil
 }
 
-func containsInput(tx *wire.MsgTx, inputHash *chainhash.Hash, inputIdx uint32) bool {
-	for _, txIn := range tx.TxIn {
-		if txIn.PreviousOutPoint.Hash.IsEqual(inputHash) && txIn.PreviousOutPoint.Index == inputIdx {
-			return true
+func findInputIdx(
+	tx *wire.MsgTx,
+	fundingTxHash *chainhash.Hash,
+	fundingOutputIdx uint32,
+) (uint32, error) {
+	for idx, txIn := range tx.TxIn {
+		if txIn.PreviousOutPoint.Hash.IsEqual(fundingTxHash) && txIn.PreviousOutPoint.Index == fundingOutputIdx {
+			return uint32(idx), nil
 		}
 	}
-	return false
+	return 0, fmt.Errorf("transaction does not spend the expected output %s:%d", fundingTxHash.String(), fundingOutputIdx)
+}
+
+func getFundingTxTransactions(txs [][]byte) ([]*wire.MsgTx, error) {
+	if len(txs) == 0 {
+		return nil, fmt.Errorf("no funding transactions provided")
+	}
+
+	fundingTxs := make([]*wire.MsgTx, len(txs))
+
+	for i, tx := range txs {
+		fundingTx, err := bbn.NewBTCTxFromBytes(tx)
+
+		if err != nil {
+			return nil, fmt.Errorf("failed to parse funding transaction: %w", err)
+		}
+
+		fundingTxs[i] = fundingTx
+	}
+
+	return fundingTxs, nil
 }
 
 // BTCUndelegate adds a signature on the unbonding tx from the BTC delegator
@@ -614,6 +638,7 @@ func (ms msgServer) BTCUndelegate(goCtx context.Context, req *types.MsgBTCUndele
 
 	btcHeader := stakerSpendigTxHeader.Header.ToBlockHeader()
 
+	// 1. Verify stake spending tx inclusion proof
 	proofValid := btcckpttypes.VerifyInclusionProof(
 		btcutil.NewTx(stakeSpendingTx),
 		&btcHeader.MerkleRoot,
@@ -625,11 +650,39 @@ func (ms msgServer) BTCUndelegate(goCtx context.Context, req *types.MsgBTCUndele
 		return nil, types.ErrInvalidBTCUndelegateReq.Wrap("stake spending tx is not included in the Bitcoin chain: invalid inclusion proof")
 	}
 
-	registeredUnbondingTx, err := bbn.NewBTCTxFromBytes(btcDel.BtcUndelegation.UnbondingTx)
+	stakingTx := btcDel.MustGetStakingTx()
+
+	stakingTxHash := stakingTx.TxHash()
+
+	// 2. Verify stake spending tx spends staking output
+	stakingTxInputIdx, err := findInputIdx(
+		stakeSpendingTx,
+		&stakingTxHash,
+		btcDel.StakingOutputIdx,
+	)
 
 	if err != nil {
-		panic(fmt.Errorf("failed to parse unbonding tx from existing delegation with hash %s: %w", req.StakingTxHash, err))
+		return nil, types.ErrInvalidBTCUndelegateReq.Wrapf("stake spending tx does not spend staking output: %s", err)
 	}
+
+	fundingTxs, err := getFundingTxTransactions(req.FundingTransactions)
+
+	if err != nil {
+		return nil, types.ErrInvalidBTCUndelegateReq.Wrapf("failed to parse funding transactions: %s", err)
+	}
+
+	// 3. Verify staker signature on stake spending tx
+	if err := VerifySpendStakeTxStakerSig(
+		btcDel.BtcPk.MustToBTCPK(),
+		stakingTx.TxOut[btcDel.StakingOutputIdx],
+		stakingTxInputIdx,
+		fundingTxs,
+		stakeSpendingTx,
+	); err != nil {
+		return nil, types.ErrInvalidBTCUndelegateReq.Wrapf("failed to verify stake spending tx staker signature: %s", err)
+	}
+
+	registeredUnbondingTx := btcDel.MustGetUnbondingTx()
 
 	registeredUnbondingTxHash := registeredUnbondingTx.TxHash()
 
@@ -648,19 +701,8 @@ func (ms msgServer) BTCUndelegate(goCtx context.Context, req *types.MsgBTCUndele
 
 		types.EmitEarlyUnbondedEvent(ctx, btcDel.MustGetStakingTxHash().String(), stakerSpendigTxHeader.Height)
 	} else {
-		// stakeSpendingTx is not unbonding tx, first we need to verify whether it
-		// actually spends staking output
-		stakingTxHash, err := chainhash.NewHashFromStr(req.StakingTxHash)
-
-		if err != nil {
-			// panic as we already verified the staking tx hash in the beginning
-			panic(fmt.Errorf("failed to parse staking tx hash from existing delegation with hash %s: %w", req.StakingTxHash, err))
-		}
-
-		if !containsInput(stakeSpendingTx, stakingTxHash, btcDel.StakingOutputIdx) {
-			return nil, types.ErrInvalidBTCUndelegateReq.Wrap("stake spending tx does not spend staking output")
-		}
-
+		// spend staking tx is not the registered unbonding tx, we need to save it in the database
+		// and emit an event
 		delegatorUnbondingInfo = &types.DelegatorUnbondingInfo{
 			SpendStakeTx: req.StakeSpendingTx,
 		}